+++ ortto_flutter_sdk_platform_interface/lib/src/models/widget_result.dart
@@ -1,0 +1,15 @@
+
+class WidgetResult {
+  bool? success;
+  String? message;
+
+  WidgetResult({
+    this.success,
+    this.message,
+  });
+
+  WidgetResult.fromMap(Map<String, dynamic> map) {
+    success = map['success'] as bool?;
+    message = map['message'] as String?;
+  }
+}